requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
omegaconf>=2.3.0
open_clip_torch>=2.24.0
PyWavelets>=1.5.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
imageio[ffmpeg]>=2.31.0
einops>=0.7.0
safetensors>=0.4.0
huggingface_hub>=0.20.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
Pillow>=10.0.0
